apps/ai/requirements.txt
opencv-python-headless
pika
psycopg2-binary
flask
numpy
websockets